requirements.txt
requests-html==0.10.0
easyimap==0.7.0
pytz==2022.1
selenium==4.1.5
pandas==1.4.2
py7zr==0.18.5
pyyaml==6.0
boto3==1.23.1
sqlalchemy==1.4.36
openpyxl==3.0.9
python-dotenv==0.20.0
pymysql==1.0.2
# pymssql==2.2.5
# pyodbc==4.0.32